soda_mmqc/tools/requirements.txt
streamlit>=1.31.0
Pillow>=10.0.0
pandas>=2.0.0 